refactor: Separar a rota de "avaliações" entre estudantes e professores
