+++ src/router/constants/routesMap.ts
@@ -2,6 +2,7 @@ const Routes = {
   HOME: '/home',
   SIGN_IN: '/sign-in',
   ASSESSMENTS: '/assessments',
+  ASSESSMENTS_STUDENT: '/assessments/student',
   NOT_FOUND: '*',
 } as const
